feature: add graceful shutdown to the jsonl inference engine
